Document enum schemas and widen match interval casts

+++ crates/sr-common/src/db/candidates.rs
@@ -118,8 +118,8 @@ pub async fn fetch_candidates_for_project(
 
             let mut response = MatchResponse {
                 interaction_id: row.get("interaction_id"),
-                talent_id: row.get::<_, i32>("talent_id") as i64,
-                project_id: row.get::<_, i32>("project_id") as i64,
+                talent_id: row.get::<_, i64>("talent_id"),
+                project_id: row.get::<_, i64>("project_id"),
                 auto_match_eligible: false,
                 manual_review_required: row.get("needs_manual_review"),
                 score: row.get::<_, Option<f64>>("score_total").unwrap_or_default() as f32,
@@ -153,8 +153,7 @@ pub async fn fetch_candidates_for_project(
                 response.manual_review_required = true;
             }
 
-            response.auto_match_eligible =
-                !is_knockout && response.is_auto_match_eligible(config);
+            response.auto_match_eligible = !is_knockout && response.is_auto_match_eligible(config);
             Ok(response)
         })
         .collect::<Result<Vec<_>, CandidateFetchError>>()?;
